PLATFORM-5639 Update Documentation to display current handling of EB06 for limitations. (limitations.time_period_qualifier)
Client pointed out that in limitations object, we parse EB06 as "time_period_qualifier", however for all other objects, we parse EB06 as "time_period"
The client is correct, but changing this now would be a breaking change for other clients.
It will have to stay as is until pokitedi V5. I'll update our documentation to reflect the way we currently display this. (limitations.time_period_qualifier)